don't highlight in isearch if ISEARCH_ACTIVE is unsupported
Old versions of zsh don't expose ISEARCH_ACTIVE. Therefore we
are unable to re-apply zle_highlight on top and it is impossible
to see the underlined area.

Completely disable highlighting in isearch in that case.

To do that, we need to make sure we are actually called when
something changes in isearch.

Trumps #257.