driver: Add a layer of indirection.
This commit copies zsh-syntax-highlighting.zsh to driver.zsh verbatim
and creates a new zsh-syntax-highlighting.zsh that consists of two
blocks from the old zsh-syntax-highlighting.zsh and a call to driver.zsh.

For the copyright claim on the new zsh-syntax-highlighting.zsh, the
year 2016 is specified as the start year, based on commits
2dce602727d715f142b5067be4f85de84d98a8b7 and
b07ada1255b74c25fbc96901f2b77dc4bd81de1a.